+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## Unreleased
+
+### Added
+
+- Custom request for client to get doxygen comment (generation of comments) for a line. The request is sent to the server and the server figures-out the the proper comment data.
+
 ## [1.3.1] - 2025-04-10
 
 ### Fixed